+++ examples/logging/src/gen/context.v1.context.rs
@@ -54,6 +54,9 @@ impl ::buffa::DefaultInstance for RequestContext {
         VALUE.get_or_init(|| ::buffa::alloc::boxed::Box::new(Self::default()))
     }
 }
+impl ::buffa::MessageFullName for RequestContext {
+    const FULL_NAME: &'static str = "buffa.examples.context.v1.RequestContext";
+}
 impl ::buffa::Message for RequestContext {
     /// Returns the total encoded size in bytes.
     ///